+++ Framework/Tween/Tween.h
@@ -66,13 +66,13 @@ public:
 	引数 type : イージングタイプ
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Move(GameObject& ref, const D3DXVECTOR3& startPosition, const D3DXVECTOR3& endPosition, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Move(GameObject& ref, const D3DXVECTOR3& startPosition, const D3DXVECTOR3& endPosition, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	移動トゥイーン
 	基本的には上記と一緒だが、移動開始座標を現在座標に自動で設定してくれる
 	***************************************/
-	static void Move(GameObject& ref, const D3DXVECTOR3& endPosition, int duratino, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Move(GameObject& ref, const D3DXVECTOR3& endPosition, float duratino, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	スケールトゥイーン
@@ -83,13 +83,13 @@ public:
 	引数 type : イージングタイプ
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Scale(GameObject& ref, const D3DXVECTOR3& startScale, const D3DXVECTOR3& endScale, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Scale(GameObject& ref, const D3DXVECTOR3& startScale, const D3DXVECTOR3& endScale, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	スケールトゥイーン
 	基本的には上記と一緒だが開始時のスケールを現在スケールに自動で設定してくれる
 	***************************************/
-	static void Scale(GameObject& ref, const D3DXVECTOR3& endScale, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Scale(GameObject& ref, const D3DXVECTOR3& endScale, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	回転トゥイーン
@@ -101,13 +101,13 @@ public:
 	引数 callback : 終了時のコールバック関数
 	※回転角度は自動で0.0f~360.0fに補正されてしまうので注意
 	***************************************/
-	static void Rotate(GameObject& ref, const D3DXVECTOR3& startEulerAngle, const D3DXVECTOR3& endEulerAngle, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Rotate(GameObject& ref, const D3DXVECTOR3& startEulerAngle, const D3DXVECTOR3& endEulerAngle, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	回転トゥイーン
 	基本的には上記と一緒だが開始の回転角度を自動で現在角度に設定してくれる
 	***************************************/
-	static void Rotate(GameObject& ref, const D3DXVECTOR3& endEulaerAngle, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Rotate(GameObject& ref, const D3DXVECTOR3& endEulaerAngle, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	方向トゥイーン
@@ -118,7 +118,7 @@ public:
 	引数 dummyAxis : 始点の向きと終点の向きが平行だった場合に使用する回転軸
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Turn(GameObject& ref, const D3DXVECTOR3& endDirection, int duration, EaseType type, const D3DXVECTOR3& dummyAxis, std::function<void()> callback = nullptr);
+	static void Turn(GameObject& ref, const D3DXVECTOR3& endDirection, float duration, EaseType type, const D3DXVECTOR3& dummyAxis, std::function<void()> callback = nullptr);
 
 	/**************************************
 	値トゥイーン
@@ -130,7 +130,7 @@ public:
 	引数 callback : 終了時のコールバック関数
 	***************************************/
 	template<class T>
-	static void To(std::shared_ptr<T>& ref, const T& start, const T& end, int duration, EaseType type, std::function<void()> callback = nullptr)
+	static void To(std::shared_ptr<T>& ref, const T& start, const T& end, float duration, EaseType type, std::function<void()> callback = nullptr)
 	{
 		ValueTweener<T> *tweener = new ValueTweener<T>(ref, start, end, duration, type, callback);
 		mInstance->tweenerContainer.push_back(tweener);
@@ -145,18 +145,18 @@ public:
 	引数 callback : 終了時のコールバック関数
 	***************************************/
 	template<class T>
-	static void To(std::shared_ptr<T>& ref, const T& end, int duration, EaseType type, std::function<void()> callback = nullptr)
+	static void To(std::shared_ptr<T>& ref, const T& end, float duration, EaseType type, std::function<void()> callback = nullptr)
 	{
 		T start = *ref;
 		ValueTweener<T> *tweener = new ValueTweener<T>(ref, start, end, duration, type, callback);
 		mInstance->tweenerContainer.push_back(tweener);
 	}
 
-	static void Expand(std::shared_ptr<Polygon2D>& ref, ExpandType expand, int duration, EaseType type, std::function<void()> callback = nullptr);
+	static void Expand(std::shared_ptr<Polygon2D>& ref, ExpandType expand, float duration, EaseType type, std::function<void()> callback = nullptr);
 
-	static void Close(std::shared_ptr<Polygon2D>& ref, CloseType close, int duration, EaseType type, std::function<void()> callback = nullptr);
+	static void Close(std::shared_ptr<Polygon2D>& ref, CloseType close, float duration, EaseType type, std::function<void()> callback = nullptr);
 
-	static void Fade(std::shared_ptr<Polygon2D>& ref, float start, float end, int duration, EaseType type, std::function<void()> callback = nullptr);
+	static void Fade(std::shared_ptr<Polygon2D>& ref, float start, float end, float duration, EaseType type, std::function<void()> callback = nullptr);
 
 private:
 	void Update();
